Use uint32 for header fields, added cmd header field

+++ read_queue.go
@@ -22,7 +22,7 @@ type readQueue struct {
 	lock           *sync.Mutex
 	cond           *sync.Cond
 	closed         *atomic.Bool
-	nextSeqId      uint16
+	nextSeqId      uint32
 	pointer        int
 }
